backend/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
falcon==2.0.0
hug==2.6.1
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
requests==2.31.0
urllib3==2.1.0
